requirements.txt
azure-mgmt-resource
azure-mgmt-network
azure-mgmt-compute
azure-cli-core